requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.8.10"
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
